Refactor
Extract more functions from playGame to shorten the code and make it more
readable.

Also, rename makeMove to addDirectionDetection and call it inside finishSetup
so that a new event listener is not added every frame.
