+++ client/src/components/game/mechanics/addDirectionDetection.js
@@ -1,4 +1,4 @@
-export default function makeMove(variables) {
+export default function addDirectionDetection(variables) {
   window.addEventListener("keydown", ({ key }) => {
     if (key === "ArrowUp") {
       variables.lastKeyPressed = "up";
