+++ client/src/components/game/mechanics/addDirectionDetection.test.js
@@ -1,8 +1,8 @@
-import makeMove from "./makeMove";
+import addDirectionDetection from "./addDirectionDetection";
 
 let mockVariables;
 
-describe("makeMove", () => {
+describe("addDirectionDetection", () => {
   beforeEach(() => {
     mockVariables = {
       lastKeyPressed: "",
@@ -11,28 +11,28 @@ describe("makeMove", () => {
 
   it("sets the key to up when the ArrowUp key is pressed", () => {
     const event = new KeyboardEvent("keydown", { key: "ArrowUp" });
-    makeMove(mockVariables);
+    addDirectionDetection(mockVariables);
     window.dispatchEvent(event);
     expect(mockVariables.lastKeyPressed).toBe("up");
   });
 
   it("sets the key to left when the ArrowLeft key is pressed", () => {
     const event = new KeyboardEvent("keydown", { key: "ArrowLeft" });
-    makeMove(mockVariables);
+    addDirectionDetection(mockVariables);
     window.dispatchEvent(event);
     expect(mockVariables.lastKeyPressed).toBe("left");
   });
 
   it("sets the key to right when the ArrowRight key is pressed", () => {
     const event = new KeyboardEvent("keydown", { key: "ArrowRight" });
-    makeMove(mockVariables);
+    addDirectionDetection(mockVariables);
     window.dispatchEvent(event);
     expect(mockVariables.lastKeyPressed).toBe("right");
   });
 
   it("sets the key to down when the ArrowDown key is pressed", () => {
     const event = new KeyboardEvent("keydown", { key: "ArrowDown" });
-    makeMove(mockVariables);
+    addDirectionDetection(mockVariables);
     window.dispatchEvent(event);
     expect(mockVariables.lastKeyPressed).toBe("down");
   });
